pyproject.toml
[tool.poetry]
name = "qqbot"
version = "0.1.0"
description = ""
authors = ["YUMEYA <shiaobo8@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.22.0"
aiobungie = {git = "https://github.com/nxtlo/aiobungie"}
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-cqhttp = "^2.0.0-beta.1"
zhconv = "^1.4.3"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true

[tool.isort]
profile = "black"
